A first complete and reasonably well working version of the "qlever" script (#1)
Supports the following actions: index, start, stop, restart, log, status, cache-stats, clear-cache, clear-cache-complete, disk-usage, memory-usage remove-index, server-settings, and ui. For each action, it also shows the (low-level) command that it executes.

The actions "index" and "start" can be used with and without docker. The default is with docker. Then no code is required on the machine except this script.

The script is supposed to be very self-explanatory. Just call it and it will guide you from there on. It's not perfect yet, but already quite usable, and of great help for many of the typical actions.