Do ulimit -Sn 1048576 before indexing input >= 10 GB
Note: ulimit -Sn sets the number of files that can be open at the same
time. It wouldn't harm to always set this number higher (the default is
1024 on a standard Ubuntu). But we want to keep the index command
simple. The 10 GB threshold is arbitray and just a proxy for "pretty
large". The input files for the complete Wikidata are around 90 GB as
ttl.bz2.