Change DOWNLOAD_URLS to DOWNLOAD_CMD and eval it
This allows more flexibility, for example, download a ZIP file and then
extracting the files (which is a rather typical use case). Added a
warning to the comment at the top of the Qleverfile that an "eval" is
happening for some of variables and that one should hence pay attention
what one writes.