Add and consider QLeverfile variable ACCESS_TOKEN
The QLever server, with PR #691, restricts access to certain URL
parameters, like cmd=clear-cache-complete or index-description=..., and
requires a correct "acess-token=..." parameter. This commit adds an
ACCESS_TOKEN variable to the default and preconfigured Qleverfiles and
sends its content to the server for the qlever script actions
"clear-cache-complete" and when setting the descriptions at the end of
the action "start".