Own docker container for actions "index" and "start"
So far, there was a single docker container, in which the actions
"index" and "start" were executed using "docker exec". This had several
disadvantages. In particular, the starting and stopping of the container
had to be managed outside of the actions. Also, there was no way to run
the server with the --restart option. Now both actions have their own docker
container and the commands are executed with "docker run".

As a side effect, remove  the actions "docker-on" and "docker-off",
which were anywhere only there as a safeguard in case somethig went
wrong with the automatic starting and stopping of the (so far single)
docker container.