Greatly simplify and improve the action "ui"
This was made possible by the commits 035e44 through 99fa52 to the
QLever UI code, which have rise to a new, clean, and now very easy to
use adfreiburg/qlever-ui docker image. In particular, there is no longer
a need to clone the qlever-ui repository or to mess around with the
database by hand, as previously done.