Use /etc/localtime in docker container
That way, the time shown in the docker container will be the same as in
the shell from where the docker container was called. So far, always UTC
time was shown, which was confusing.

NOTE: It seems that -v /etc/localtime:/etc/localtime:ro is enough. The
typical recommendation is to also have -v /etc/timezone:/etc/timezon:ro,
but that didn't seem to be necessary.