Improve index-stats action wrt text index info
The text index is sometimes build after the normal index. The action
`qlever index-stats` now considers this and shows the correct time
for the text index build and the total even if there is time gap between
the end of the build of the normal index and beginning of the build of
the text index.