First version of script with basic actions
Actions supported: get-data, index, start, stop, status

The start action already supports USE_DOCKER=true (stop and index don't
yet).

It's not yet complete, but already works for olympics on an Ubuntu 22.04
machine, with the QLever binaries `ServerMain` and `IndexBuilderMain` in
the path.

The associated `Qleverfile.ini` is also committed. Note that eventually,
this will be called `Qleverfile` again (without `.ini`), but for now we
want to distinguish it from the original, so that we can have both in
the same directory.