Add action "index-stats" + qlever.py -> qlever
There is now also an action "index-stats" like in the old script, just
more robust and with better error messages when something goes wrong.
Also test that action in all the workflows.

The new script is now called `qlever` (and not qlever.py), and the old
script is called `qlever.OLD`. Correspondingly, the new Qleverfiles are
now just called `Qleverfile`, while the old script now produces and
expects `Qleverfile.OLD`.