Major refactoring of the qlever script
1. Switch to using argparse and argcomplete with commands, subparsers,
   and normal `--` options. That is more standard then the current
   syntax.

2. So far, this was a monolithic Python script. There is now one `.py`
   file for each command, in a dedicated folder `commands`. Everything
   specific for the command (including a short help text, which
   arguments from the Qleverfile are relevant for it, and which
   additional arguments there are) is in the respective `.py` file. All
   the generic code is elsewhere. That will make it very easy to add new
   commands.

This is still work in progress. Two commands are already there (`index`
and `sehr_doof`). Neither does something meaningful yet (that will be
the easy part=, but the argument parsing and overrding and
autocompletion works (that was the hard part). The Qleverfile is not yet
read (that will be easy to add).