First working version of the config parser
This was the hard part. In particular, autocompletion works now in a
pretty generic way. And the help info shows which arguments have values
in the Qleverfile.

TODO: Right now, the help info is still Python-style. At some point,
we should format this in a way that is more consistent with the output
of the rest of the script.