Implement commands `setup-config`, `get-data`, `index`
Works really well and already significantly more powerful than before,
in particular:

1. It's now easy to add additional options like --overwrite-existing for
   the index command
2. There is now autocompletion for eyerything, also values of options
   with choices (e.g., for `with_text_index` or `runtime_environment`)
3. Now supports three runtime environments: `docker`, `podman`,
   `native` ... and it was straightforward to implement
4. Several other improvements in the old code

The code is also much cleaner than before because all the generic code
is in proper modules and the module for a command really only contains
the code needed for that command and only has access to what it needs
access to. I am quite happy with how this turned out (after much
thinking about the best design, but that was worth it).