Change default UI port for QLever UI to 8176 (#55)
The default for the QLever UI was 7000, but this port is not available on macOS (which is used by a lot of our users). We now take port 8176, which happens to be the concatenated ASCII codes of the first two letters of QLever and is *not* listed on https://en.wikipedia.org/wiki/List_of_TCP_and_UDP_port_numbers as a widely used port (neither officially nor unofficially). Fixes #49