Warn if the UI port is already in use (#56)
When port `UI_PORT` is already in use, the command `qlever ui` failed with a long and technical docker error message. Now try to detect before if the port is available and if not, print a short and easily comprehensible error message with instructions on how to change `UI_PORT`. Fixes #49