Add Qlevefile variable FORMAT and support NQ
So far, only NT and TTL were supported, which both used the same
(Turtle) parser, so `IndexBuilderMain` was always called with option `-F
ttl`. Since we are now approaching support of named graphs, add a config
variable FORMAT with the options nt, ttl, nq, and which will be the
argument of the `-F` option.