Add Qlevefile variable FORMAT and support NQ (#59)
So far, the only input formats supported were NTriples and Turtle. In QLever, they both use the same (Turtle) parser, so the `qlever` script always called `IndexBuilderMain` with option `-F ttl`. Since we are now approaching support of named graphs, add a config variable `FORMAT` with the options `nt`, `ttl`, *and* `nq`, which then become the argument of the `-F` option in the call to `IndexBuilderMain`.