Improve Qleverfile for PubChem
The `get-data` command now does not only download the latest version of
the PubChem data, but also of the associated ontologies. That is
relatively little additional data but makes a big difference for
usability. See the NOTEs in the new `Qleverfile.pubchem` for a longer
explanation.

TODO: Still testing whether this works as intended.