Much improved Qleverfile for PubChem (#60)
The `get-data` command now does not only download the latest version of the PubChem data, but also of the associated ontologies. That is relatively little additional data but makes a big difference for usability. Also remove the fixing of the invalid IRIs in the PubChem data (using `sed` and `awk`, which wasn't cheap) and instead rely on the new https://github.com/ad-freiburg/qlever/pull/1448, which activates more relaxed parsing of IRI refs by default. I tried it and it worked like a charm. See the beginning of the new `Qleverfile.pubchem` for the resource requirements and a longer explanation.