Only stop index container if one is actually running (#70)
So far, `qlever index` always called `docker rm -f ...` or `podman rm -f ...` before starting a new index build. However, old versions of podman return an error message when no container with the specified name is running. This is now fixed by checking explicitly whether a container is running and only if there is, trying to stop and remove it. Fixes #66.