Better statistics for `example-queries` command (#75)
Now show TOTAL, AVERAGE, and MEDIAN, and also the number of FAILED queries in case there were any. Along the way, improve the layout and the code.

Also, for the first time, use `black` for autoformatting the code. A pull request, where I will do this for all `.py` files, will follow soon.