New variable `MULTI_INPUT_JSON` for Qleverfile
So far, each Qleverfile needed to have a value for `CAT_INPUT_FILES`,
which specified a command that wrote the input to stdout (which was then
piped into `IndexBuilderMain`).

This is still possible, but there is also an alternative variable
`MULTI_INPUT_JSON` for reading multiple streams via multiple pipes. The
streams are read concurrently, and for each stream, a format and a graph
can be specified separately, as well as whether that stream should be
parsed in parallel. See `qlever index --help` for more information.

The Qleverfile for Wikidata now uses `MULTI_INPUT_JSON`.