New variable `MULTI_INPUT_JSON` for Qleverfile (#77)
So far, the input to QLever's index builder was always single input stream. Correspondingly, each Qleverfile had a variable `CAT_INPUT_FILES`, which specifies a command that writes the input to standard ouput, which is then piped into the index builder

Since https://github.com/ad-freiburg/qlever/pull/1537, QLever allows multipe input streams. The streams are parsed concurrently. For each stream separately, the following can be specified: the command that writes the input to standard output, the format, the graph of which this input should become a part, and whether the stream should be parsed in parallel (recommended for very large inputs but then all the prefix declarations must come at the beginning) or not.

In the Qleverfile, multiple input streams and their configuration can now be specifed via `MULTI_INPUT_JSON`, which takes a JSON string as value . See `qlever index --help` or `src/qlever/qleverfile.py` for information about the structure of that JSON. Using `CAT_INPUT_FILES` is still supported, but it has to be either `MULTI_INPUT_JSON` or `CAT_INPUT_FILES`. 

The Qleverfile for Wikidata has been updated to use `MULTI_INPUT_JSON` (as a showcase and because it makes sense).