Update Qleverfile for OHM Planet
Add options `--add-way-node-order` and `--add-way-node-geometry` to
`osm2rdf`. This adds the (extensive) complete information on which ways
contain which nodes and the geometries of those nodes, even if they are
used for nothing else and have no tags.

Also clean up the Qleverfile and update the information at the top.