New Qleverfile for PubChem using `MULTI_INPUT_JSON`
Since https://github.com/ad-freiburg/qlever/pull/1537 QLever supports
multiple input streams, with an individual configuration per stream.

This is now used for PubChem, where it is convenient to assign a graph
not for each triple individually, but based on the directory of the
input file.

TODO: This version provides a coarse partition of the input by
directory. The indexing would be fast if we partition the input by file.