For the `MULTI_INPUT_JSON` value, also allow `JSONL` (#91)
So far, `MULTI_INPUT_JSON` had to be a JSON string. When producing the string programatically for a large number of directories or files, that can be cumbersome because of the leading `[` and trailing `]` and the separating commas.

To make this easier, the string may now also be in JSONL format (one JSON per line). It will then simply be transformed to a JSON array in the obvious way.

As a proof of concept and as an example, the Qleverfile for DBLP is adapted accordingly.