Easy `MULTI_INPUT_JSON` specification with `for-each`
So far, specifying the `MULTI_INPUT_JSON` was cumbersome when the input
consists of many files, assigned to few graphs via a pattern. For
example, when all files in a directory should be assigned to the same
graph. There is now a nice syntax for this typical case, here is an
example element of the `MULTI_INPUT_JSON`, which should be self-explanatory:

{ "cmd": "zcat {}", "graph": "http://example.org", "for-each": "ex/*.gz" }

Update the `Qleverfile` for UniProt to use this new feature (and read
the input as gzipped TTL files, and not N-Triples as before).