If command was not successful, exit with code 1
So far, the exit code of the `qlever` script was 1 only when the command
threw an exception that was not caught by the command. This happens when
there is something wrong with the command itself or a certain error case
is not handled.

Now the exit code is 1 also when the command returns `False`, which
signals that the command was not successful. This is important when
executing chains of commands, for example

`qlever index && qlever start`

where the second command should not be executed if the first one failed.