Refactor the `start` command and add unit tests (#69)
Previously, the `start` command had a very long `execute` function which as of this PR now is refactored into smaller functions 
(e.g. for handling the native or containerized starting and for killing preexisting binaries or containers with the same port.
Also add unit tests (using mock tests) for the `start` command.