New `extract-queries` command + other improvements (#115)
The new `extract-queries` command extracts all queries from the server log and writes it into a TSV file that can be processed using the `example-queries` command. If the server log contains lines matching `Alive check with message "..."`, the message is used as base name for the query descriptions. For each query with that base name `Query #<index>` is appended to the description.

In this context, the following small improvements for the `example-queries` command (items 1, 2, 3, 4) and the `settings` command (item 5) were made:

1. The default for `--download_or_count` is now the more natural `download` (was: `count`).
2. The option `--accept` now has two more choices `application/qlever-results+json` and `AUTO`, where the later picks the accept header automatically from the query type
3. There is a new option `--add-query-type-to-description`
4. Long query descriptions are now truncated in the middle, that is shown as `Beginnig ... end`
5. For the runtime parameters supported by the settings command, there is now also `always-mulitply-unions`