Remove `void.ttl.gz` from `Qleverfile.uniprot` (#104)
The `void.ttl.gz` that is provided from the UniProt data distribution refers to the division of the data by file name. For the SPARQL endpoint, we want statistics by graph name. Also, the additional data (Rhea and Chebi)  should be considered. So the way to go is to generate the VoID data from the SPARQL endpoint (for example, via a tool like https://github.com/JervenBolleman/void-generator) and then add it in a dedicated graph https://sparql.uniprot.org/.well-known/void