Prepare the `qlever` script for support for other engines (#154)
Originally, the `qlever` command-line tool was written for easy access to the functionality provided by QLever. As it turns out, many of the commands are also useful for other engines and it would be very convenient if the script could be used for other engines with the same ease that it can currently be used with QLever. This change refactors the script (slightly) to prepare it for such an extension. There should be no change in functionality.