Let `qlever ui` control QLever UI configuration via YAML file (#141)
With this change, the `qlever ui` command allows control of the QLever UI configuration via a new `Qleverfile-ui.yml` file. Specifically:

1. The command checks if `Qleverfile-ui.yml` exists in the current directory
2. If yes, that file is used to set the QLever UI configuration accordingly
3. If not, the command does the following:
 a. It takes the value of `UI_CONFIG` from the normal Qleverfile (`default` by default)
 b. It looks for a default config with that name in the QLever UI image (`adfreiburg/qlever-ui` by default)
 c. It writes that config to `Qleverfile-ui.yml`

Once, `Qleverfile-ui.yml` is there, a user my simply edit it and then run `qlever ui` again. The command will check and report if there is something wrong with the file. The file also contains the full configuration for the autocompletion queries and all the example queries for that particular backend. The file does not contain any user-specific information (except maybe in the `baseUrl`) and can be exchanged between different users.

Addresses https://github.com/ad-freiburg/qlever-ui/issues/125

Co-authored-by: Julian Mundhahs <munhahj@cs.uni-freiburg.de>