Proper aggregation of messages into single update operation
The triples of each message are now properly parsed (using `rdflib`) and
properly aggregated into a single update operation per group. In
particular, if a triple occurs multiple times in messages of one group,
only the last operation (be it an insert or delete) is considered. This
is always correct, no matter whether the triple is already ontained
in the graph or not.

On the side, further improve logging and error handling.

For example, with a group size of 100,000 messages and a sufficiently
large `request-body-limit` (`1G` is enough, better choose `10G`), the
`qlever update` command now runs through without any errorr, at a rate
of approximately 50 µs per operation.