Further improve statistics
This now takes into account that the JSON returned by QLever is an
array. In particular, `cache-stats` and `settings` are adapated to
handle that case (they were expecting a single object so far).

NOTE: This requires `fix/timingInformationExperiment` by `Qup42`, which
provides fine-grained timing information that is used for the statistics
shown by the `update` command.