Finish batch when approaching NOW, support Ctrl+C
The former "groups" are now called "batches". A batch is now
automatically finished when a message is encountered that is
within `lag-seconds` of the current time (default: 1 second).

Also a batch can now be finished by pressing Ctrl+C. The program
will then finish assembling the current batch, process it, and end
gracefully. In order to continue processing, just press Ctrl+C again.

Due to these changes, the `--num-batches` option has been removed
and the default for the `--batch-size` option has been changed to
100,000 (which is a good size for when messages are still far away from
the current time).