Command `update-wikidata` (#158)
Command for updating a Wikidata SPARQL endpoint via the messages of the official Wikidata update SSE stream: https://stream.wikimedia.org/v2/stream/rdf-streaming-updater.mutation.v2 . The command can be used very easily as follows, from the directory with the `Qleverfile` for the endpoint being updated:
```
qlever update-wikidata
```
The `since` parameter of the update stream (that is, the date starting from which messages are consumed) is set automatically via a hard-coded SPARQL query. That query returns the latest `wikibase:Dump wikibase:updatesCompleteUntil ...` date (these triples are added by the update requests issued by this command), or, if none of these triples exist yet, the earliest `wikibase:Dump schema:dateModified ...` date (these triples are added during the creation of the dataset dump). That way, the command automatically resumes from the right date, no matter whether it is started for the first time for a given endpoint, or whether it was aborted (for any reason) and then restarted again.

The command reads the following fields from each message of the stream:
```
`dt`: the date of the message
`rdf_added_data`: triples to be inserted, assumed to be in TTL or NT format
`rdf_deleted_data`: triples to be deleted, assumed to be in TTL or NT format
```
The messages are batched and one update request is issued for each batch. A batch is finished when 100'000 messages have been collected (this number can be controlled via `--batch-size`), or when a message is encountered that is less than 1 second behind the present moment (this lag can be controlled via `--lag-seconds`). When a batch is finished for the latter reason, the command waits 3600s before processing the next batch (this time can be controlled via `--wait-between-batches`). For more arguments and help, see `qlever update-wikidata --help`.